products/requirements.txt
dj-rest-auth==2.2.5
Django==4.0.5
django-allauth==0.51.0
django-bootstrap3==21.2
django-cors-headers==3.2.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0